expand_Rscript {fundoc}	R Documentation

File path handler

Description

File path handler

Usage

expand_Rscript(fps, debug = F)

Arguments

fps
A character vector of paths. The vector can be a mix of paths of either (1) directory paths, (2) file paths, or (3) glob expressions to R scripts. Note that (2) and (3) must end with .R or .r.

debug
Logical. Whether to return results as a list with fps as names.

Value

A character vector or a list.

Examples

expand_Rscript = fundoc:::expand_Rscript
dir_ = system.file(package = "fundoc")
(a_file = file.path(dir_, "foo.R"))

## [1] "C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/foo.R"

expand_Rscript(a_file)

## NULL

(a_dir = file.path(dir_, "R"))

## [1] "C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/R"

expand_Rscript(a_dir)

## character(0)

(a_glob = file.path(dir_, "*", "*.R"))

## [1] "C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/*/*.R"

expand_Rscript(a_glob)

## character(0)

# Mix file paths
expand_Rscript( c(a_file, a_dir, a_glob) )

## character(0)

expand_Rscript( c(a_file, a_dir, a_glob), debug = T )

## $`C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/foo.R`
## NULL
## 
## $`C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/R`
## character(0)
## 
## $`C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/*/*.R`
## character(0)

[Package fundoc version 0.0.1 Index]