fundoc {fundoc}	R Documentation

Generate documentation from R scripts

Description

Generate documentation from R scripts

Usage

fundoc(
  scripts = ".",
  out_dir = "docs",
  proj_name = NULL,
  proj_title = NULL,
  version = NULL,
  quiet = FALSE
)

Arguments

scripts
A character vector of paths. The vector can be a mix of paths of either (1) directory paths, (2) file paths, or (3) glob expressions to R scripts. Note that (2) and (3) must end with .R or .r. See expand_Rscript() for details.

proj_name
Character. The name of the project. Must conform to conventions of naming R packages. To check whether a name is valid, refer to https://github.com/r-lib/available. Defaults to NULL, which generates the name Foo.{YYYY.MM.DD} according to the current date.

proj_title
Character. A one-sentence title of what the project does. Defaults to NULL, which generates a title of the format: Function documentation of project {proj_name}.

version
Character. Version number of the project. Defaults to "0.0.1" when set to NULL.

quiet
Logical. Whether to hide messages from the subprocess compiling the function documentation. Defaults to FALSE.

outdir
Path to the generated documentation directory.

Examples

pkg = system.file(package="fundoc")
(scripts = file.path(pkg, "*.R"))

## [1] "C:/Users/rd/AppData/Local/Temp/RtmpEPXiaG/INST/fundoc/*.R"

## Not run: 
##D fundoc(scripts,
##D        out_dir = "docs/reference/docs",
##D        proj_name = "fundoc",
##D        proj_title = "Quick and Dirty Function Documentation")
## End(Not run)

[Package fundoc version 0.0.1 Index]