locate_description {fundoc}	R Documentation

Locate DESCRIPTION content from a bunch of files

Description

Given a vector of file paths, this function tries to find and parse the R script with DESCRIPTION content written as Roxygen docstring at the beginning of the file. If more than one file contain DESCRIPTION content, only the first one would be parsed and returned.

Usage

locate_description(files)

Arguments

files
A character vector of file paths.

[Package fundoc version 0.0.1 Index]